+++ .gitignore
@@ -1,2 +1,3 @@
 cache.txt
 .env
+releases*
