+++ env.example
@@ -1,2 +1,3 @@
 export QOTD_CACHE_FILE "$PWD/cache.txt"
 export QOTD_URL "http://quotes.rest/qod.json?category=management"
+export DOCKERPROD="productionDockerAddr"
