+++ templates/index.html
@@ -3,21 +3,14 @@
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width" />
-    <title>Quotes</title>
-    <link rel="stylesheet" href="https://bootswatch.com/sandstone/bootstrap.min.css" />
-    <style type="text/css" media="all">
-      body {
-        background: url("{{.BG}}");
-        background-size: 100em;
-        background-repeat: no-repeat;
-        background-position: center;
-      }
-    </style>
+    <title>Morning Inspiration</title>
+    <link rel="stylesheet" href="/static/bootstrap.min.css"/>
+    <link rel="stylesheet" href="/static/stylesheet.css"/>
   </head>
   <body>
     <div class="container">
       <div class="row">
-        <div class="col-md-6">
+        <div class="col-md-6 col-md-offset-3">
           <div class="well well-lg">
             <h1>{{.Author}}</h1>
             <blockquote>
