+++ templates/static/stylesheet.css
@@ -1,0 +1,10 @@
+/* body { */
+/*   [> background-image: url("/static/background.jpg"); <] */
+/*   background-size: cover; */
+/*   background-repeat: no-repeat; */
+/*   background-position: center; */
+/* } */
+
+.container {
+  margin-top: 2em;
+}
